requirements.txt
vapi_server_sdk
fastapi[standard]
openai

